Added Pan For Gold 2.0.0
This is the version of Pan for Gold that is actually being worked on. As you can see it is a lot better than the last one because it actually uses methods. If you're reading this here's a few things that I intend to add in the future: A shop with a variety of containers for carrying acid to purify gold, and a variety of pans that multiply the amount of gold you get, possibly an inventory system for individual chunks of gold, with varying purity and weights, with a possibility for differing purifying methods, and possible consequences for improper practices.